clean (some) gdb warning: round 1 * use svnrev.h on linux too * replace sprintf_s with snprintf (hope it still compile on Windows) * init integer with 0 instead of NULL * various int -> u32/uint32/uint on for loop index * remove a couple of unused variable * init few variable * disable unused warning results
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@5683 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/onepad/Linux/dialog.cpp
@@ -412,8 +412,8 @@ void joy_changed(GtkComboBox *box, gpointer user_data)
 	// 3 : use new joy with fresh key binding
 
 	// unassign every joystick with this pad
-	for (int i = 0; i < 2; ++i)
-		if (joyid == conf->get_joyid(i)) conf->set_joyid(i, -1);
+	for (u32 i = 0; i < 2; ++i)
+		if (joyid == (int)conf->get_joyid(i)) conf->set_joyid(i, -1);
 
 	conf->set_joyid(current_pad, joyid);
 }
